Move the length abstraction out of ast into util and use std::gcd
len_abs does not depend on expressions or on anything else in seq, so it does
not belong in ast and does not need the seq namespace or the seq_ file prefix.
Move src/ast/seq_len_abs.{h,cpp} to src/util/len_abs.{h,cpp} and drop the
namespace, so the class is simply len_abs. The unit test moves with it and is
registered as tst_len_abs.

Replace the hand written u_gcd with std::gcd. The lcm helper stays local and is
renamed bounded_lcm: std::lcm cannot be used because its behaviour is undefined
exactly on the overflow that the helper has to detect and report as "degrade to
the gcd instead".

No functional change: the header and body are otherwise identical to the
originals, and the 13 benchmarks this branch newly decides are unaffected.

Copilot-Session: a2ce3573-4e15-4a4a-afb5-21e3cb04e4a2
