+++ src/test/CMakeLists.txt
@@ -79,6 +79,7 @@ add_executable(test-z3
   interval.cpp
   karr.cpp
   lcube.cpp
+  len_abs.cpp
   list.cpp
   main.cpp
   map.cpp
@@ -131,7 +132,6 @@ add_executable(test-z3
   sat_user_scope.cpp
   scoped_timer.cpp
   scoped_vector.cpp
-  seq_len_abs.cpp
   seq_rewriter.cpp
   seq_monadic.cpp
   seq_monadic_bench.cpp
